Bardzo duże zmiany na bazie, update liquibase, encji, dto i tak dalej. Pewnie będzie sporo błędów w JSF, musicie je sobie poprawić. Polecam dropnięcie i postawienie na nowo bazy. Pozdro

+++ project/project-core/src/main/java/pl/lodz/p/project/core/domain/document/base/InvoiceGood.java
@@ -7,26 +7,25 @@ import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.math.BigDecimal;
 
 /**
  * @author Jan Krajewski
  */
 @MappedSuperclass
-public class InvoiceGood<T extends Serializable> extends BaseEntity<T>  {
+public class InvoiceGood<T extends Serializable> extends BaseEntity<T> {
 
     private static final long serialVersionUID = 4953126124258975050L;
 
     @Basic(optional = false)
     @NotNull
     @Column(name = "quantity")
-    private BigDecimal quantity;
+    private Double quantity;
 
-    public BigDecimal getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(BigDecimal quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
     }
